Refactor person properties update code so it can be reused
